refactor python difference calculation
